requirements.txt
# Core dependencies
PyYAML>=6.0.1
aiohttp>=3.8.5
aiofiles>=23.1.0
typing-extensions>=4.7.1

# Speech processing
pyttsx3>=2.90
vosk>=0.3.44
pyaudio>=0.2.13
pvporcupine>=2.2.1
numpy>=1.24.0

# Intent parsing
adapt-parser>=0.4.0
dateparser>=1.1.8  # For parsing dates and times

# Google Calendar API
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.95.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Development tools
black>=23.7.0
isort>=5.12.0  # For import sorting
mypy>=1.4.1   # For type checking